+++ Example/Tests/freestyle/Styling/PXStylesheetParserTests.m
@@ -463,7 +463,7 @@
     /*PXStylesheet *stylesheet =*/ [parser parse:source withOrigin:PXStylesheetOriginApplication];
 
     NSArray *errors = parser.errors;
-    XCTAssertTrue(errors.count == 0, @"Found %d unexpected parse error(s): %@", errors.count, [errors componentsJoinedByString:@"\n"]);
+    XCTAssertTrue(errors.count == 0, @"Found %lu unexpected parse error(s): %@", (unsigned long)errors.count, [errors componentsJoinedByString:@"\n"]);
 }
 
 - (void)testNthOfType
@@ -473,7 +473,7 @@
     /*PXStylesheet *stylesheet =*/ [parser parse:source withOrigin:PXStylesheetOriginApplication];
 
     NSArray *errors = parser.errors;
-    XCTAssertTrue(errors.count == 0, @"Found %d unexpected parse error(s): %@", errors.count, [errors componentsJoinedByString:@"\n"]);
+    XCTAssertTrue(errors.count == 0, @"Found %lu unexpected parse error(s): %@", (unsigned long)errors.count, [errors componentsJoinedByString:@"\n"]);
 }
 
 - (void)testNthChildWithoutSpaces
@@ -483,7 +483,7 @@
     /*PXStylesheet *stylesheet =*/ [parser parse:source withOrigin:PXStylesheetOriginApplication];
 
     NSArray *errors = parser.errors;
-    XCTAssertTrue(errors.count == 0, @"Found %d unexpected parse error(s): %@", errors.count, [errors componentsJoinedByString:@"\n"]);
+    XCTAssertTrue(errors.count == 0, @"Found %lu unexpected parse error(s): %@", (unsigned long)errors.count, [errors componentsJoinedByString:@"\n"]);
 }
 
 - (void)testNthChildWithoutSpaces2
@@ -493,7 +493,7 @@
     /*PXStylesheet *stylesheet =*/ [parser parse:source withOrigin:PXStylesheetOriginApplication];
 
     NSArray *errors = parser.errors;
-    XCTAssertTrue(errors.count == 0, @"Found %d unexpected parse error(s): %@", errors.count, [errors componentsJoinedByString:@"\n"]);
+    XCTAssertTrue(errors.count == 0, @"Found %lu unexpected parse error(s): %@", (unsigned long)errors.count, [errors componentsJoinedByString:@"\n"]);
 }
 
 - (void)testNthChildWithSpaces
@@ -503,7 +503,7 @@
     /*PXStylesheet *stylesheet =*/ [parser parse:source withOrigin:PXStylesheetOriginApplication];
 
     NSArray *errors = parser.errors;
-    XCTAssertTrue(errors.count == 0, @"Found %d unexpected parse error(s): %@", errors.count, [errors componentsJoinedByString:@"\n"]);
+    XCTAssertTrue(errors.count == 0, @"Found %lu unexpected parse error(s): %@", (unsigned long)errors.count, [errors componentsJoinedByString:@"\n"]);
 }
 
 - (void)testNthLastChild
@@ -513,7 +513,7 @@
     /*PXStylesheet *stylesheet =*/ [parser parse:source withOrigin:PXStylesheetOriginApplication];
 
     NSArray *errors = parser.errors;
-    XCTAssertTrue(errors.count == 0, @"Found %d unexpected parse error(s): %@", errors.count, [errors componentsJoinedByString:@"\n"]);
+    XCTAssertTrue(errors.count == 0, @"Found %lu unexpected parse error(s): %@", (unsigned long)errors.count, [errors componentsJoinedByString:@"\n"]);
 }
 
 - (void)testNthLastOfType
@@ -523,7 +523,7 @@
     /*PXStylesheet *stylesheet =*/ [parser parse:source withOrigin:PXStylesheetOriginApplication];
 
     NSArray *errors = parser.errors;
-    XCTAssertTrue(errors.count == 0, @"Found %d unexpected parse error(s): %@", errors.count, [errors componentsJoinedByString:@"\n"]);
+    XCTAssertTrue(errors.count == 0, @"Found %lu unexpected parse error(s): %@", (unsigned long)errors.count, [errors componentsJoinedByString:@"\n"]);
 }
 
 - (void)testImportant
@@ -557,7 +557,7 @@
     /*PXStylesheet *stylesheet =*/ [parser parse:source withOrigin:PXStylesheetOriginApplication];
 
     NSArray *errors = parser.errors;
-    XCTAssertTrue(errors.count == 0, @"Found %d unexpected parse error(s): %@", errors.count, [errors componentsJoinedByString:@"\n"]);
+    XCTAssertTrue(errors.count == 0, @"Found %lu unexpected parse error(s): %@", (unsigned long)errors.count, [errors componentsJoinedByString:@"\n"]);
 }
 
 - (void)testImportWithURL
@@ -567,7 +567,7 @@
     /*PXStylesheet *stylesheet =*/ [parser parse:source withOrigin:PXStylesheetOriginApplication];
 
     NSArray *errors = parser.errors;
-    XCTAssertTrue(errors.count == 0, @"Found %d unexpected parse error(s): %@", errors.count, [errors componentsJoinedByString:@"\n"]);
+    XCTAssertTrue(errors.count == 0, @"Found %lu unexpected parse error(s): %@", (unsigned long)errors.count, [errors componentsJoinedByString:@"\n"]);
 }
 
 #pragma mark - Keyframes Tests
@@ -582,7 +582,7 @@
     XCTAssertNotNil(keyframe, @"Expected a 'test' keyframe to be defined in the stylesheet");
 
     NSArray *blocks = keyframe.blocks;
-    XCTAssertTrue(blocks.count == 0, @"Expected no keyframe blocks but found %d", blocks.count);
+    XCTAssertTrue(blocks.count == 0, @"Expected no keyframe blocks but found %lu", (unsigned long)blocks.count);
 }
 
 - (void)testKeyframesWithTo
@@ -595,11 +595,11 @@
     XCTAssertNotNil(keyframe, @"Expected a 'test' keyframe to be defined in the stylesheet");
 
     NSArray *blocks = keyframe.blocks;
-    XCTAssertTrue(blocks.count == 1, @"Expected 1 keyframe block but found %d", blocks.count);
+    XCTAssertTrue(blocks.count == 1, @"Expected 1 keyframe block but found %lu", (unsigned long)blocks.count);
 
     PXKeyframeBlock *block = blocks[0];
     XCTAssertTrue(block.offset == 1.0f, @"Expected offset to be 1.0, but found %f", block.offset);
-    XCTAssertTrue(block.declarations.count == 1, @"Expected 1 declaration but found %d", block.declarations.count);
+    XCTAssertTrue(block.declarations.count == 1, @"Expected 1 declaration but found %lu", (unsigned long)block.declarations.count);
 }
 
 - (void)testKeyframesWithFrom
@@ -612,11 +612,11 @@
     XCTAssertNotNil(keyframe, @"Expected a 'test' keyframe to be defined in the stylesheet");
 
     NSArray *blocks = keyframe.blocks;
-    XCTAssertTrue(blocks.count == 1, @"Expected 1 keyframe block but found %d", blocks.count);
+    XCTAssertTrue(blocks.count == 1, @"Expected 1 keyframe block but found %lu", (unsigned long)blocks.count);
 
     PXKeyframeBlock *block = blocks[0];
     XCTAssertTrue(block.offset == 0.0f, @"Expected offset to be 0.0, but found %f", block.offset);
-    XCTAssertTrue(block.declarations.count == 1, @"Expected 1 declaration but found %d", block.declarations.count);
+    XCTAssertTrue(block.declarations.count == 1, @"Expected 1 declaration but found %lu", (unsigned long)block.declarations.count);
 }
 
 - (void)testKeyframesWithPercentage
@@ -629,11 +629,11 @@
     XCTAssertNotNil(keyframe, @"Expected a 'test' keyframe to be defined in the stylesheet");
 
     NSArray *blocks = keyframe.blocks;
-    XCTAssertTrue(blocks.count == 1, @"Expected 1 keyframe block but found %d", blocks.count);
+    XCTAssertTrue(blocks.count == 1, @"Expected 1 keyframe block but found %lu", (unsigned long)blocks.count);
 
     PXKeyframeBlock *block = blocks[0];
     XCTAssertTrue(block.offset == 0.5f, @"Expected offset to be 0.5, but found %f", block.offset);
-    XCTAssertTrue(block.declarations.count == 1, @"Expected 1 declaration but found %d", block.declarations.count);
+    XCTAssertTrue(block.declarations.count == 1, @"Expected 1 declaration but found %lu", (unsigned long)block.declarations.count);
 }
 
 - (void)testKeyframesWithMultipleBlocks
@@ -646,19 +646,19 @@
     XCTAssertNotNil(keyframe, @"Expected a 'test' keyframe to be defined in the stylesheet");
 
     NSArray *blocks = keyframe.blocks;
-    XCTAssertTrue(blocks.count == 3, @"Expected 3 keyframe blocks but found %d", blocks.count);
+    XCTAssertTrue(blocks.count == 3, @"Expected 3 keyframe blocks but found %lu", (unsigned long)blocks.count);
 
     PXKeyframeBlock *block = blocks[0];
     XCTAssertTrue(block.offset == 0.0f, @"Expected offset to be 0.0, but found %f", block.offset);
-    XCTAssertTrue(block.declarations.count == 1, @"Expected 1 declaration but found %d", block.declarations.count);
+    XCTAssertTrue(block.declarations.count == 1, @"Expected 1 declaration but found %lu", (unsigned long)block.declarations.count);
 
     block = blocks[1];
     XCTAssertTrue(block.offset == 0.5f, @"Expected offset to be 0.5, but found %f", block.offset);
-    XCTAssertTrue(block.declarations.count == 1, @"Expected 1 declaration but found %d", block.declarations.count);
+    XCTAssertTrue(block.declarations.count == 1, @"Expected 1 declaration but found %lu", (unsigned long)block.declarations.count);
 
     block = blocks[2];
     XCTAssertTrue(block.offset == 1.0f, @"Expected offset to be 1.0, but found %f", block.offset);
-    XCTAssertTrue(block.declarations.count == 1, @"Expected 1 declaration but found %d", block.declarations.count);
+    XCTAssertTrue(block.declarations.count == 1, @"Expected 1 declaration but found %lu", (unsigned long)block.declarations.count);
 }
 
 - (void)testKeyframesWithMultipleOffsets
@@ -671,15 +671,15 @@
     XCTAssertNotNil(keyframe, @"Expected a 'test' keyframe to be defined in the stylesheet");
 
     NSArray *blocks = keyframe.blocks;
-    XCTAssertTrue(blocks.count == 2, @"Expected 2 keyframe blocks but found %d", blocks.count);
+    XCTAssertTrue(blocks.count == 2, @"Expected 2 keyframe blocks but found %lu", (unsigned long)blocks.count);
 
     PXKeyframeBlock *block = blocks[0];
     XCTAssertTrue(block.offset == 0.0f, @"Expected offset to be 0.0, but found %f", block.offset);
-    XCTAssertTrue(block.declarations.count == 1, @"Expected 1 declaration but found %d", block.declarations.count);
+    XCTAssertTrue(block.declarations.count == 1, @"Expected 1 declaration but found %lu", (unsigned long)block.declarations.count);
 
     block = blocks[1];
     XCTAssertTrue(block.offset == 1.0f, @"Expected offset to be 1.0, but found %f", block.offset);
-    XCTAssertTrue(block.declarations.count == 1, @"Expected 1 declaration but found %d", block.declarations.count);
+    XCTAssertTrue(block.declarations.count == 1, @"Expected 1 declaration but found %lu", (unsigned long)block.declarations.count);
 }
 
 #pragma mark - @media Tests
